pyproject.toml
[project]
name = "SAADA"
version = "0.1.0"
description = "基于多智能体的 AI 系统态势感知与自动化防御系统"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # --- LLM & Agent 框架 ---
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-google-genai>=4.2.0",
    "langgraph>=0.2.0",
    "google-adk>=1.22.1",

    # --- 数据验证 ---
    "pydantic>=2.0.0",

    # --- HTTP & 解析 ---
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",

    # --- 配置 ---
    "python-dotenv>=1.0.0",

    # --- API 服务 ---
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
saads = "main:main"
